requirements.txt
flask>=2.3.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0
numpy>=1.24.0
reportlab>=4.0.0
# 可选依赖(根据需要安装)
# aliyun-python-sdk-core>=2.13.0  # 阿里云短信
# tencentcloud-sdk-python>=3.0.0  # 腾讯云短信
# twilio>=8.0.0  # Twilio国际短信
